fixing delete slider bug

+++ app/Http/Controllers/Admin/SliderController.php
@@ -150,7 +150,9 @@ class SliderController extends Controller
     public function destroy($id)
     {
         $slider = Slider::find($id);
-        unlink('uploads/slider/'.$slider->image);
+        if(file_exists('uploads/slider/'.$slider->image)) {
+            unlink('uploads/slider/'.$slider->image);
+        }
         $slider->delete();
         return redirect()->back()->with('success', 'Slider Successfully Deleted');
     }
